refactor(cli): centralize getCleanedRewindText logic
Moves duplicated text cleaning logic into a shared utility file (formatters.ts) to ensure consistency between interactive and non-interactive rewind paths.
